refactor: extract localStorage helpers into separate file
- Create feature-opt-in-storage.ts for localStorage helper functions
- Update useFormbricksOptInTracking to import from new storage file
- Update useFeatureOptInBanner to import from new storage file
